+++ src/components/children/DistributionTable.jsx
@@ -33,25 +33,53 @@ class DistributionTable extends React.Component {
 						</TableCell>
 					</TableRow>
 					<TableRow>
-						<TableCell>Limit State</TableCell>
-						<TableCell>Alpha/
-							{
-								curves[0].alphaType?
-									curves[0].alphaType[0].toUpperCase()
-								+ curves[0].alphaType.slice(1)
-									:
-									""
-							}</TableCell>
-						<TableCell>Beta</TableCell>
+						{
+							Object.keys(curves[0]).map(key =>{
+								if (key === "description") return (<TableCell>limit state</TableCell>);
+								else if (key !== "className") return (<TableCell>{key}</TableCell>);
+							})
+						}
 					</TableRow>
 				</TableHead>
 				<TableBody displayRowCheckbox={false}>
 					{curves.map(function (curve) {
 						return (
 							<TableRow>
-								<TableCell>{curve.description}</TableCell>
-								<TableCell>{curve.alpha}</TableCell>
-								<TableCell>{curve.beta}</TableCell>
+								{
+									Object.keys(curve).map(key =>{
+										if ((key === "alpha" || key === "beta") && curve.className === "ConditionalStandardFragilityCurve"){
+											return (<TableCell>
+												{
+													curve[key].map(content => {
+														return (<TableRow>{content}</TableRow>);
+													})
+												}
+											</TableCell>);
+										}
+										else if (key === "rules" && curve.className === "ConditionalStandardFragilityCurve"){
+											return (<TableCell>
+												{
+													curve.alpha.map((a, index) => {
+														return (<TableRow>{(curve[key][index.toString()])}</TableRow>);
+													})
+												}
+											</TableCell>);
+										}
+										else if (key === "parameters" && curve.className === "ParametricFragilityCurve"){
+											return (<TableCell>
+												{
+													curve[key].map(parameterSet => {
+														return (
+															<TableRow>{JSON.stringify(parameterSet)}</TableRow>);
+													})
+												}
+											</TableCell>);
+										}
+										else if (key !== "className") {
+											return (<TableCell>{curve[key]}</TableCell>);
+										}
+									})
+								}
 							</TableRow>
 						);
 					})}