+++ src/utils/chartSampler.js
@@ -18,6 +18,37 @@ export default class chartSampler {
 		return samples;
 	}
 
+	static computeParametricSampes(min, max, numberOfSamples, curveType, parameters){
+		let steps = ((max - min) / numberOfSamples);
+		let samples = [];
+
+		if (curveType.toLowerCase() === "logit"){
+			for (let i = 1; i <= numberOfSamples; i++) {
+				let A1 = 1; // coefficient for demand (X)
+				let cumulateTerm = 0; // X * Theta'
+				for (let j = 0; j< parameters.length; j++){
+					let name = parameters[j]["name"].toLowerCase();
+					let interceptTermDefault = parameters[j]["interceptTermDefault"];
+					let coefficient = parameters[j]["coefficient"];
+
+					if (name === "demand"){
+						A1 = coefficient;
+					}
+					else{
+						cumulateTerm += interceptTermDefault * coefficient;
+					}
+				}
+
+				let y = steps * i;
+				let x = math.exp((math.log(y/(1-y)) - cumulateTerm) / A1);
+				samples.push([x, y]);
+			}
+		}
+		// TODO: else if other curve type
+
+		return samples;
+	}
+
 	static async computeExpressionSamples3d(minX, maxX, numberOfSamplesX, minY, maxY, numberOfSamplesY, expression) {
 		let stepsX = ((maxX - minX) / numberOfSamplesX);
 		let stepsY = ((maxY - minY) / numberOfSamplesY);
@@ -60,6 +91,19 @@ export default class chartSampler {
 		return samples;
 	}
 
+	static sampleConditional(min, max, numberOfSamples, alphaType, rules, alpha, beta){
+		let samples = [];
+
+		if (alphaType === "median"){
+			samples = chartSampler.sampleLogNormalCdfConditional(min, max, numberOfSamples, rules, alpha.map(a => Math.log(a)), beta);
+		}
+		else if (alphaType === "lambda"){
+			samples = chartSampler.sampleLogNormalCdfConditional(min, max, numberOfSamples, rules, alpha, beta);
+		}
+
+		return samples;
+	}
+
 	static sampleNormalCdf(min, max, numberOfSamples, mean, std) {
 		let steps = ((max - min) / numberOfSamples);
 
@@ -87,6 +131,48 @@ export default class chartSampler {
 			let y = jStat.lognormal.inv(x, mean, std);
 
 			samples.push([y, x]);
+		}
+
+		return samples;
+
+	}
+
+	static sampleLogNormalCdfConditional(min, max, numberOfSamples, rules, mean_arr, std_arr) {
+		const knownOperator = {
+			"EQ": "==",
+			"EQUALS": "==",
+			"NEQUALS": "!=",
+			"GT": ">",
+			"GE": ">=",
+			"LT": "<",
+			"LE": "<=",
+		};
+
+		let steps = ((max - min) / numberOfSamples);
+
+		let samples = [];
+
+		for (let key in Object.keys(rules)){
+
+			let mean = mean_arr[parseInt(key)];
+			let std = std_arr[parseInt(key)];
+			let elements = rules[key][0].split(" ", 3);
+			let ruleOperator = elements[1];
+			let ruleValue = elements[2];
+			if (!(ruleOperator in knownOperator)){
+				// console.log( ruleOperator + " Unknown. Cannot parse the rules!");
+				return samples;
+			}
+
+			for (let i = 1; i <= numberOfSamples; i++) {
+				let x = steps * i;
+				let y = jStat.lognormal.inv(x, mean, std);
+
+				// if match the rules, put that dot in the curve; otherwise disgard
+				if (eval(y.toString() + knownOperator[ruleOperator] + ruleValue)){
+					samples.push([y, x]);
+				}
+			}
 		}
 
 		return samples;